+++ src/meson.build
@@ -14,11 +14,11 @@ keyframe_sources = [
 	'timeline/keyframe-timeline-property.c',
 	'timeline/keyframe-timeline-header.c',
 	'timeline/keyframe-timeline-channel.c',
-	'timeline/keyframe-track.c',
-	'timeline/keyframe-track-clip.c',
-	'timeline/keyframe-track-clip-layout.c',
-	'timeline/keyframe-track-clip-widget.c',
-	'timeline/keyframe-track-frame.c',
+	'timeline/tracks/keyframe-track.c',
+	'timeline/tracks/keyframe-track-clip.c',
+	'timeline/tracks/keyframe-track-clip-layout.c',
+	'timeline/tracks/keyframe-track-clip-widget.c',
+	'timeline/tracks/keyframe-track-frame.c',
 	'model/keyframe-layer.c',
 	'model/keyframe-layer-geometry.c',
 	'model/keyframe-layer-text.c',
