+++ src/timeline/tracks/keyframe-track-clip-layout.c
