+++ src/timeline/tracks/keyframe-track-clip-widget.c
