+++ src/timeline/tracks/keyframe-track-clip.h
