+++ src/timeline/tracks/keyframe-track-frame.h
